+++ app/src/main/AndroidManifest.xml
@@ -44,7 +44,7 @@
         </receiver>
 
         <service
-            android:name=".AppInstallerService"
+            android:name=".appinstaller.AppInstallerService"
             android:enabled="true"
             android:exported="false"></service>
     </application>
